app/requirements.txt
Django==5.0.6
psycopg2-binary==2.9.9
moviepy==1.0.3
pillow==10.3.0